restructuring of tests
