+++ src/test/java/com/rcoem/filmrentalstore/api/LanguageApiTest.java
@@ -1,4 +1,4 @@
-package com.rcoem.filmrentalstore;
+package com.rcoem.filmrentalstore.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
